requirements.txt
Django==4.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-bootstrap5==23.3
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
coverage==7.3.2
pylint==3.0.2